dashboard/requirements.txt
streamlit
mysql-connector-python
psycopg2-binary
pika
pandas